refactor(celery): remove RedisClientProtocol and datamodels, consolidate into single module
- celery_reliability.py: delete RedisClientProtocol(Protocol) and replace
  with `Redis` imported from `app.utils.cache` (re-exported from redis.asyncio,
  same type used by cache/ and app.connections.redis). Update all
  `redis_client: Redis` signatures and `ResilientTask.get_redis_client() -> Redis`.
- Remove Pydantic datamodels from this file: `IdempotencyRecord(BaseModel)` and
  `CircuitBreakerState(BaseModel)` replaced with plain dict + `to_json_str`/
  `from_json` (wire format unchanged: {"status","task_id","updated_at","metadata"}),
  `CircuitBreakerSnapshot` as `TypedDict`. Keep `RateLimitResult` as `BaseModel`
  per clarification (was @dataclass, now frozen BaseModel) — no dataclass in file.
- Keep `RateLimiter` + `ReliabilitySystem` + `idempotency_manager` as plain
  functions/classes; no @dataclass left in file.
- Back-compat: `celery_reliability.__getattr__` re-creates `CircuitBreakerState`,
  `IdempotencyRecord`, `RedisClientProtocol` as lazy shims so
  `from app.connections.celery_reliability import CircuitBreakerState` keeps
  working for one release (tests).
- Consolidate `app/connections/celery*` into single source `celery.py`:
  `celery_task_names.py` stays as single definition site (imported by celery.py),
  `celery_reliability.py` and `celery_registry.py` become thin shims re-exporting
  from `celery.py` (`from app.connections.celery import ...`). New code must
  import from `app.connections.celery` directly.
- Fix missing `app/shared/langchain_layer/agents/memory/setup_types.py`
  (CogneeSetupConfig) that broke `real_celery` import.
- Fix test harness: remove `app.connections.celery` mock from
  `tests/conftest.py` and `tests/unit/celery/test_reliability_system.py`
  circular-mock (now that celery is single source), lift
  `app.connections.celery_reliability` in `tests/unit/celery/conftest.py`.

Fixes #44
Closes #226
